+++ scanner/scanner/identifiers/guess/guess.py
